pyproject.toml
[tool.poetry]
name = "mptt-dropbox"
version = "0.1.0"
description = ""
authors = ["Iris Faey Hoffmeyer <hi@iris.best>"]
license = "mit"

[tool.poetry.dependencies]
python = "^3.8"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
